Add SIZE_T for HDF5
It turns out that SIZE_T from HDF5 can be 8-byte, which can be different
from Fortran integer. This mismatch can lead to a HDF5 issue. In the
previous commit, some SIZE_T are introduced to remove type warnings, but
they were not complete. This commit fixes this issue.
It is likely to be more of these data type issues lurking somewhere, and
they will be addressed later.

+++ dsm2/src/hdf_tidefile/hdf5_write_attributes.f90
@@ -68,8 +68,8 @@ subroutine hdf5_write_attributes()
     integer(HSIZE_T), dimension(2) :: h_offset
     integer(HID_T) :: memspace ! memspace identifier
 
-    integer, parameter :: label_len = 12
-    integer, parameter :: name_len = 32
+    integer(SIZE_T), parameter :: label_len = 12
+    integer(SIZE_T), parameter :: name_len = 32
     character(LEN=label_len), dimension(2) :: chan_location &
                                               = (/"upstream", "downstream"/)
     character(LEN=name_len), dimension(:), allocatable :: names
